+++ lib/shared/infrastructure/acp_admin/acp_admin_repository_impl.dart
@@ -32,6 +32,7 @@ class AcpAdminRepositoryImpl implements AcpAdminRepository {
         queryParameters: <String, dynamic>{
           r'$top': top,
           r'$orderby': 'Name asc',
+          r'$filter': "Status eq 'active'",
         },
       ),
     );